Moved the cluster specific components under cluster folder
