--- src/Components/ModulesList.test.js
@@ -1,8 +1,0 @@
-import { mount } from 'enzyme';
-import ModulesList from './ModulesList';
-
-describe('Components/ModulesList', () => {
-  it('should render successfully', () => {
-    mount(<ModulesList modules={[{ name: 'Foo', host_task_count: 1 }]} />);
-  });
-});
